requirements.txt
python-telegram-bot
qrcode
python-dotenv
markdown
pytz
markdown2pdf
reportlab